refactor: reduce dynamic routes
