+++ app/(admin)/layout.tsx
@@ -1,5 +1,6 @@
 import { AppSidebar } from "@/components/app-sidebar";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
+import AuthorizedApolloWrapper from "@/providers/use-apollo.rsc";
 import ProtectedRoute from "@/providers/use-protected-route";
 import { unstable_ViewTransition as ViewTransition } from "react";
 
@@ -10,21 +11,23 @@ export default function AdminLayout({
 }>) {
   return (
     <ProtectedRoute>
-      <SidebarProvider
-        style={{
-          "--sidebar-width": "calc(var(--spacing) * 72)",
-          "--header-height": "calc(var(--spacing) * 12)",
-        } as React.CSSProperties}
-      >
-        <AppSidebar variant="inset" />
-        <SidebarInset>
-          <ViewTransition>
-            <div suppressHydrationWarning>
-              {children}
-            </div>
-          </ViewTransition>
-        </SidebarInset>
-      </SidebarProvider>
+      <AuthorizedApolloWrapper>
+        <SidebarProvider
+          style={{
+            "--sidebar-width": "calc(var(--spacing) * 72)",
+            "--header-height": "calc(var(--spacing) * 12)",
+          } as React.CSSProperties}
+        >
+          <AppSidebar variant="inset" />
+          <SidebarInset>
+            <ViewTransition>
+              <div suppressHydrationWarning>
+                {children}
+              </div>
+            </ViewTransition>
+          </SidebarInset>
+        </SidebarProvider>
+      </AuthorizedApolloWrapper>
     </ProtectedRoute>
   );
 }
