+++ app/forbidden.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from "next";
 
 export const metadata: Metadata = {
   title: "權限不足",
-}
+};
 
 export default function ForbiddenPage() {
   return <ForbiddenLayout />;
