+++ providers/use-apollo.rsc.tsx
@@ -1,0 +1,14 @@
+"use server";
+
+import { getAuthToken } from "@/lib/auth";
+import { ApolloWrapper } from "./use-apollo";
+
+export default async function AuthorizedApolloWrapper({ children }: { children: React.ReactNode }) {
+  const token = await getAuthToken();
+
+  return (
+    <ApolloWrapper token={token}>
+      {children}
+    </ApolloWrapper>
+  );
+}
